308310 Prepare for very large arrays

+++ plugins/org.eclipse.mat.api/src/org/eclipse/mat/snapshot/inspections/DominatorQuery.java
@@ -177,7 +177,7 @@ public class DominatorQuery implements IQuery
         int objectId;
 
         String label;
-        int shallowHeap;
+        long shallowHeap;
         long retainedHeap;
 
         public Node(int objectId)
